backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
mysql-connector-python==9.3.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
